% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatePriority.r
\name{AnnotatePriority}
\alias{AnnotatePriority}
\title{Annotation priorization based on genomic feature}
\usage{
AnnotatePriority(TableDMRAnnoIn, NumProc = 1)
}
\arguments{
\item{TableDMRAnnoIn}{Is the output from \code{PoreMethAnnotate}, data.frame.}

\item{NumProc}{Number of threads. Default = 1, numeric.}
}
\value{
data.frame with Prioritized Annotation.
}
\description{
Returns a annotation priorizatized by genomic feature
}
