% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BidimensionalSLMSegIn.r
\name{BidimensionalSLMSegIn}
\alias{BidimensionalSLMSegIn}
\title{Truncated Gaussian Bidimensional SLM Segmentation}
\usage{
BidimensionalSLMSegIn(DataMatrix, muk, mi, smu, sepsilon, omega, eta)
}
\description{
Truncated Gaussian Bidimensional SLM Segmentation
}
