% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonodimensionalSLMSegIn.r
\name{MonodimensionalSLMSegIn}
\alias{MonodimensionalSLMSegIn}
\title{Truncated Gaussian Monodimensional SLM Segmentation}
\usage{
MonodimensionalSLMSegIn(
  DataMatrix,
  muk,
  mi,
  smu,
  sepsilon,
  Pos,
  omega,
  eta,
  stepeta
)
}
\description{
Truncated Gaussian Monodimensional SLM Segmentation
}
