% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreMeth1DMRStatistics.r
\name{PoreMeth1DMRStatistics}
\alias{PoreMeth1DMRStatistics}
\title{Functions for Calculating the size of DMR in each genic feature}
\usage{
PoreMeth1DMRStatistics(
  TableInDMR,
  Assembly = "hg19",
  BetaThr = 0.2,
  PValueThr = 0.05
)
}
\arguments{
\item{Assembly}{,}

\item{BetaThr}{numeric,}

\item{PValueThr}{numeric,}
}
\value{
what what what what what what what what what
}
\description{
description description description description description description
}
