% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreMeth2PairedExpQualityPlot.r
\name{PoreMeth2PairedExpQualityPlot}
\alias{PoreMeth2PairedExpQualityPlot}
\title{Display \code{beta_cov} and \code{entropy_cov} stats for paired samples.}
\usage{
PoreMeth2PairedExpQualityPlot(TableTestIn, TableControlIn)
}
\arguments{
\item{TableTestIn}{Output from \code{parse_nanopolish_entropy.pl}, data.frame.}

\item{TableControlIn}{Output from \code{parse_nanopolish_entropy.pl}, data.frame.}
}
\value{
Plot(s)
}
\description{
With this function it is possible to display stats about \code{beta_cov} and \code{entropy_cov} for common CpG dineucleotides in the Test/Control samples pair.
}
