% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreMeth2SingleExpQualityPlot.r
\name{PoreMeth2SingleExpQualityPlot}
\alias{PoreMeth2SingleExpQualityPlot}
\title{Functions for printing plots for the evaluation of input data quality}
\usage{
PoreMeth2SingleExpQualityPlot(TableIn)
}
\arguments{
\item{TableIn}{Output from \code{parse_nanopolish_entropy.pl}}
}
\value{
Plot(s)
}
\description{
This function prints plots for the evaluation of input data quality by displaying stats about \code{beta_cov} and \code{entropy_cov} (see Data Preparation) and beta and S density functions.
}
