% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreMeth2StateFinder.r
\name{PoreMeth2StateFinder}
\alias{PoreMeth2StateFinder}
\title{Function for HMM segmentation}
\usage{
PoreMeth2StateFinder(SignalIn, PosIn, muk, sepsilon, NormDist, PTStart)
}
\arguments{
\item{PosIn}{numeric,}

\item{muk}{numeric,}

\item{sepsilon}{numeric,}

\item{NormDist}{numeric,}

\item{PTStart}{numeric,}
}
\value{
what what what what what what what what what
}
\description{
description description description description description description
}
