% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreMethAnnotate.r
\name{PoreMethAnnotate2}
\alias{PoreMethAnnotate2}
\title{Main Function for DMRs Annotation}
\usage{
PoreMethAnnotate2(
  TableMethIn,
  FileOut,
  NumProc = 5,
  AnnotationType = "Genes",
  Assembly = "hg19"
)
}
\arguments{
\item{TableMethIn}{Output from PoreMeth2DMR (or PoreMeth1DMR), data.frame.}

\item{NumProc}{Optional argument for the number of cores to use in parallel, numeric.}

\item{AnnotationType}{Optional argument that specifies whether to annotate DMRs on genic elements only (\code{"Genes"}) or genic elements and regulatory features (\code{"GenesReg"}), character.}

\item{Assembly}{Optional parameter that specifies the reference version to use for annotation (\code{"hg19"} or \code{"hg38"}), character.}
}
\value{
data.frame with DMRs annotated to genomic and regulatory elements.
}
\description{
Function for genomic and regulatory annotation of results from PoreMeth2DMR or PoreMeth1DMR
}
