% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreMethState.r
\name{PoreMethState}
\alias{PoreMethState}
\title{Function to identify methylation state from Beta and Entropy Values}
\usage{
PoreMethState(TableIn, PTStart = 0.1, ratiosd = 0.1, NormDist = 1e+06, FW = 2)
}
\arguments{
\item{TableIn}{}

\item{PTStart}{numeric,}

\item{ratiosd}{numeric,}

\item{NormDist}{numeric,}

\item{FW}{numeric,}
}
\value{
what what what what what what what what what
}
\description{
description description description description description description
}
