% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionAnnotateCompact.r
\name{RegionAnnotateCompact}
\alias{RegionAnnotateCompact}
\title{Parallel Function for Compact Annotation}
\usage{
RegionAnnotateCompact(
  j,
  ChrMethInUSel,
  StartMethInC,
  EndMethInC,
  MatrixMethResultsC,
  CoordGenCodeC,
  CoordCGIC,
  NameCGIC,
  CoordEnhancerC,
  NameEnhancerC,
  CoordDNASEC,
  NameDNASEC,
  CoordTFBSC,
  NameTFBSC,
  FeatureGenCodeC,
  StrandGenCodeC,
  SymbolGenCodeC,
  TypeGenCodeC,
  AnnotationType
)
}
\description{
Parallel Function for Compact Annotation
}
