% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionAnnotatePriority.r
\name{RegionAnnotatePriority}
\alias{RegionAnnotatePriority}
\title{Annotation Priority}
\usage{
RegionAnnotatePriority(
  Input,
  TableDMRAnnoIn,
  ChrStartEndString,
  ChrStartEndStringU,
  indFinal
)
}
\description{
Annotation Priority
}
