% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegResults.r
\name{SegResults}
\alias{SegResults}
\title{Function to calculate the mean value of each segment}
\usage{
SegResults(DataSeq, TotalPredBreak)
}
\description{
Function to calculate the mean value of each segment
}
